// ui/src/lib/pocketbase-types.ts
export enum Collections {
	Authorigins = "_authOrigins",
	Externalauths = "_externalAuths",
	Mfas = "_mfas",
	Otps = "_otps",
	Superusers = "_superusers",
	Hops = "hops",
	Searches = "searches",
	Users = "users",
}

// Alias types for improved usability
export type IsoDateString = string
export type RecordIdString = string
export type HTMLString = string

// System fields
export type BaseSystemFields<T = never> = {
	id: RecordIdString
	collectionId: string
	collectionName: Collections
	expand?: T
}
// ...
export type HopsRecord<Theaders = unknown, Tipinfo = unknown> = {
	body?: HTMLString
	created?: IsoDateString
	headers?: null | Theaders
	hop_number: number
	host?: string
	http_version?: string
	id: string
	ipaddr?: string
	ipinfo?: null | Tipinfo
	path?: string
	scheme?: string
	search_id?: RecordIdString
	status_code: number
	status_phrase?: string
	time_elapsed?: number
	updated?: IsoDateString
	url: string
}
// ...
export type HopsResponse<Theaders = unknown, Tipinfo = unknown, Texpand = unknown> = Required<HopsRecord<Theaders, Tipinfo>> & BaseSystemFields<Texpand>